api-gateway/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
